+++ includes/Derivatives.h
@@ -3,7 +3,7 @@
 class Grid;
 
 
-inline double safe_dx(double dx) {
+inline float safe_dx(float dx) {
     return (dx > 1e-12) ? dx : 1e-12;  // Évite les divisions instables
 }
 
@@ -11,20 +11,20 @@ inline double safe_dx(double dx) {
 
 template <class T>
 __attribute__((always_inline))
-inline double fourth_order_diff(const T &plus2, const T &plus1, 
-                                const T &minus1, const T &minus2, double dx)
+inline float fourth_order_diff(const T &plus2, const T &plus1, 
+                                const T &minus1, const T &minus2, float dx)
 {
     dx = safe_dx(dx);
-    double t1 = std::fma(8.0, plus1, -plus2);   
-    double t2 = std::fma(-8.0, minus1, minus2); 
-    double numerator = t1 + t2;
+    float t1 = std::fma(8.0, plus1, -plus2);   
+    float t2 = std::fma(-8.0, minus1, minus2); 
+    float numerator = t1 + t2;
     return numerator / (12.0 * dx);
 }
 
 
 template <class T>
 __attribute__((always_inline))
-inline double second_order_diff(const T &plus1, const T &minus1, double dx)
+inline float second_order_diff(const T &plus1, const T &minus1, float dx)
 {
     return (plus1 - minus1)/(2.0*dx);
 }
@@ -39,35 +39,35 @@ class Derivatives
         friend class GridTensor;
         Grid* grid;
 
-        double partialX_alpha(int i, int j, int k);
-        double partialY_alpha(int i, int j, int k);
-        double partialZ_alpha(int i, int j, int k);
+        float partialX_alpha(int i, int j, int k);
+        float partialY_alpha(int i, int j, int k);
+        float partialZ_alpha(int i, int j, int k);
 
-        double partialXX_alpha(int i, int j, int k);
-        double partialYY_alpha(int i, int j, int k);
-        double partialZZ_alpha(int i, int j, int k);
+        float partialXX_alpha(int i, int j, int k);
+        float partialYY_alpha(int i, int j, int k);
+        float partialZZ_alpha(int i, int j, int k);
 
-        double partialX_betacomp(int i, int j, int k, int comp);
-        double partialY_betacomp(int i, int j, int k, int comp);
-        double partialZ_betacomp(int i, int j, int k, int comp);
+        float partialX_betacomp(int i, int j, int k, int comp);
+        float partialY_betacomp(int i, int j, int k, int comp);
+        float partialZ_betacomp(int i, int j, int k, int comp);
 
-        double partialXY_alpha(int i, int j, int k);
-        double partialXZ_alpha(int i, int j, int k);
-        double partialYZ_alpha(int i, int j, int k);
+        float partialXY_alpha(int i, int j, int k);
+        float partialXZ_alpha(int i, int j, int k);
+        float partialYZ_alpha(int i, int j, int k);
 
-		double second_partial_alpha(int i, int j, int k, int a, int b);
+		float second_partial_alpha(int i, int j, int k, int a, int b);
 };
 
 
-void spectral_derivative_3D(const std::vector<double> &f_in,
-				std::vector<double> &f_out,
+void spectral_derivative_3D(const std::vector<float> &f_in,
+				std::vector<float> &f_out,
 				int N_x, int N_y, int N_z,
-				double dx, double dy, double dz,
+				float dx, float dy, float dz,
 				int dim);
 
 
 template <typename Func>
-double partialX(Grid &grid_obj, int i, int j, int k, Func f) {
+float partialX(Grid &grid_obj, int i, int j, int k, Func f) {
 	if (i < 0 || i >= NX) return 0.0;
 	if (j < 0 || j >= NY) return 0.0;
 	if (k < 0 || k >= NZ) return 0.0;
@@ -95,7 +95,7 @@ double partialX(Grid &grid_obj, int i, int j, int k, Func f) {
 
 
 template <typename Func>
-double partialY(Grid &grid_obj, int i, int j, int k, Func f) {
+float partialY(Grid &grid_obj, int i, int j, int k, Func f) {
 	if (j < 0 || j >= NY) return 0.0;
 	if (i < 0 || i >= NX) return 0.0;
 	if (k < 0 || k >= NZ) return 0.0;
@@ -122,7 +122,7 @@ double partialY(Grid &grid_obj, int i, int j, int k, Func f) {
 }
 
 template <typename Func>
-double partialZ(Grid &grid_obj, int i, int j, int k, Func f) {
+float partialZ(Grid &grid_obj, int i, int j, int k, Func f) {
 	if (k < 0 || k >= NZ) return 0.0;
 	if (i < 0 || i >= NX) return 0.0;
 	if (j < 0 || j >= NY) return 0.0;
@@ -150,7 +150,7 @@ double partialZ(Grid &grid_obj, int i, int j, int k, Func f) {
 
 
 template <typename Getter>
-double partial_m(Grid &grid_obj, int i, int j, int k, int dim, Getter getter) {
+float partial_m(Grid &grid_obj, int i, int j, int k, int dim, Getter getter) {
     if (dim == 0) return partialX(grid_obj, i, j, k, getter);
     if (dim == 1) return partialY(grid_obj, i, j, k, getter);
     if (dim == 2) return partialZ(grid_obj, i, j, k, getter);
@@ -158,7 +158,7 @@ double partial_m(Grid &grid_obj, int i, int j, int k, int dim, Getter getter) {
 }
 
 template <typename Getter>
-double second_partial(Grid &grid_obj, int i, int j, int k, int a, int b, Getter getter) {
+float second_partial(Grid &grid_obj, int i, int j, int k, int a, int b, Getter getter) {
     if (a == b) {
         if (a == 0 && i >= 1 && i <= NX - 2) {
             return (getter(grid_obj.getCell(i+1, j, k)) - 2.0 * getter(grid_obj.getCell(i, j, k)) + getter(grid_obj.getCell(i-1, j, k))) / (DX * DX);
@@ -177,8 +177,8 @@ double second_partial(Grid &grid_obj, int i, int j, int k, int a, int b, Getter 
         if (b == 1) { jp = j + 1; jm = j - 1; }
         if (b == 2) { kp = k + 1; km = k - 1; }
 
-        double dx_a = (a == 0) ? DX : (a == 1) ? DY : DZ;
-        double dx_b = (b == 0) ? DX : (b == 1) ? DY : DZ;
+        float dx_a = (a == 0) ? DX : (a == 1) ? DY : DZ;
+        float dx_b = (b == 0) ? DX : (b == 1) ? DY : DZ;
 
         dx_a = safe_dx(dx_a);
         dx_b = safe_dx(dx_b);
